# ModuleManagementClient/src/app/core/modules/openapi/model/module-version-compact-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ModuleVersionCompactDTO { 
    moduleVersionId?: number;
    version?: number;
    status?: ModuleVersionCompactDTO.StatusEnum;
}
export namespace ModuleVersionCompactDTO {
    export type StatusEnum = 'PENDING_SUBMISSION' | 'PENDING_FEEDBACK' | 'ACCEPTED' | 'REJECTED' | 'OBSOLETE' | 'CANCELLED';
    export const StatusEnum = {
        PendingSubmission: 'PENDING_SUBMISSION' as StatusEnum,
        PendingFeedback: 'PENDING_FEEDBACK' as StatusEnum,
        Accepted: 'ACCEPTED' as StatusEnum,
        Rejected: 'REJECTED' as StatusEnum,
        Obsolete: 'OBSOLETE' as StatusEnum,
        Cancelled: 'CANCELLED' as StatusEnum
    };
}